requirements.txt
pytest == 7.4.3
allure-pytest == 2.13.2
requests
python-dotenv
jsonschema ==4.20.0